iOS continue....

+++ unidbg-ios/src/main/native/ios/hook/objc.h
@@ -7,4 +7,4 @@ void new_objc_msgSend(id self, SEL _cmd);
 
 uintptr_t pre_objc_msgSend(id self, SEL _cmd, va_list args);
 
-typedef void (*objc_msgSend_callback)(bool systemClass, const char *className, const char* cmd, long lr);
+typedef void (*objc_msgSend_callback)(bool systemClass, const char *className, const char* cmd, uintptr_t lr);
